remove operators for SGFCNodeTraits and SGFCPropertyTraits
unlike in C++, enumerations in Objective-C are not true types, they are
merely typedefs with an underlying primitive type.

because of this the operators in question trigger a compiler error
when compiled with a C++ compiler:

  overloaded 'operator&' must have at least one parameter of class or enumeration type

the operators are unnecessary anyway since, enums not being true types,
it is possible to perform logical bitwise operations on enums just like
that.